final
changed the home css and home.jsx
