+++ src/App.js
@@ -6,7 +6,7 @@
 // import PrivateRoute from "/Users/pranaysinguluri/movie-bluff/src/Components /PrivateRoute.jsx";
 
 import { Routes, Route} from "react-router-dom";
-import Homly from "../src/Pages/Homly.jsx";
+import Home from "./Pages/Home.jsx";
 import Login from "../src/Pages/Login.jsx";
 import SignUp from "../src/Pages/SignUp.jsx";
 import About from "../src/Pages/About.jsx";
@@ -15,13 +15,15 @@ import Plot from "../src/Pages/Plot.jsx";
 import ErrorPage from "../src/Pages/ErrorPage.jsx";
 import Profile from "../src/Pages/Profile.jsx";
 import Admin from "../src/Pages/Admin.jsx";
+
 // import {Logout }from "/Users/pranaysinguluri/movie-bluff/src/Pages/Logout.jsx";
 
 function App() {
+
+
   return (
     <Routes>
-      <Route path="*" element={<Homly />} />
-      <Route path="/home/*" element={<Homly />} />
+      <Route path="*" element={<Home />} />
       <Route path="/login" element={<Login />} />
       <Route path="/signup" element={<SignUp />} />
       <Route path="/about" element={<About/>}/>
