+++ src/Pages/Home.jsx
@@ -1,210 +1,198 @@
-// import React, { useState } from "react";
-// import { Routes, Route, Link, useLocation } from "react-router-dom";
-// import Search from "../Components /SearchBar.jsx";
-// import Plot from "../Pages/Plot.jsx";
-// import About from "../Pages/About.jsx";
-// import NavBar from "../Components /NavBar.jsx";
-// import LoadingGif from "../assets/loadingGif.gif";
-// import ErrorPage from "../Pages/ErrorPage.jsx";
-// import { VscStarHalf } from "react-icons/vsc";
-// import Footer from "../Components /Footer.jsx";
-// import Login from "./Login";
-// import SignUp from "../Pages/SignUp.jsx";
-// import Upcoming from "../Components /Trending.jsx";
-// import useFetch from "../Hooks/useFetch.jsx";
-// import TrendingMovies from "../Components /Trending.jsx";
-// import Welcome from "./Welcome.jsx";
-// // import token from "/Users/pranaysinguluri/movie-bluff/src/Utilities/token.jsx"
-
-// const TMDB_BASE_URL = "https://api.themoviedb.org/3";
-
-// function Home() {
-//   const [url, setUrl] = useState("");
-//   const [searchTerm, setSearchTerm] = useState("");
-//   const { data: movies, loading, error } = useFetch(url);
-
-//   const location = useLocation();
-//   const userName = location.state?.username;
-
-//   const handleSearch = (query) => {
-//     setSearchTerm(query);
-//     setUrl(`${TMDB_BASE_URL}/search/movie?query=${query}`);
-//   };
-
-//   return (
-//     <>
-//       <NavBar />
-//       welcome to Movie Bluff {userName}
-//       <Routes>
-//         <Route
-//           path="/home"
-//           element={
-//             <div className="wrapper-search">
-//               <Search onSearch={handleSearch} />
-
-//               {loading && (
-//                 <div className="loading-container">
-//                   <img src={LoadingGif} alt="Loading..." width="200" />
-//                 </div>
-//               )}
-
-//               {error && <p className="error-message">{error}</p>}
-
-//               {searchTerm && movies && movies.length > 0 && (
-//                 <p className="results-info">
-//                   {`Results for "${searchTerm}": ${movies.length}`}
-//                 </p>
-//               )}
-
-//               {!searchTerm || (movies && movies.length === 0) ? <Upcoming /> : null} 
-//               {/* shows upcoming  */}
-
-//               {movies && movies.length > 0 && (
-//                 <div className="movie-grid">
-//                   {movies.map((movie) => (
-//                     <div key={movie.id} className="movie-card">
-//                       <h2>{movie.title}</h2>
-//                       <p>{movie.release_date?.split("-")[0]}</p>
-//                       <p>Language: {movie.original_language.toUpperCase()}</p>
-//                       <p>
-//                         <VscStarHalf /> {movie.vote_average.toFixed(1)}
-//                       </p>
-
-//                       <Link to={`/plot?id=${movie.id}`}>
-//                         <img
-//                           src={`https://image.tmdb.org/t/p/w500/${movie.poster_path}`}
-//                           alt={movie.title}
-//                           className="movie-poster"
-//                         />
-//                       </Link>
-//                     </div>
-//                   ))}
-//                 </div>
-//               )}
-//               <Footer />
-//             </div>
-//           }
-//         />
-//         <Route path="/plot" element={<Plot />} />
-//         <Route path="/welcome" element={<Welcome/>} />
-//         <Route path="/trending" element={<TrendingMovies />} />
-//         <Route path="/about" element={<About />} />
-//         <Route path="/login" element={<Login />} />
-//         <Route path="/signup" element={<SignUp />} />
-//         <Route path="/*" element={<ErrorPage />} />
-//       </Routes>
-//     </>
-//   );
-// }
-
-// export default Home;
-
-
 import React, { useState, useEffect } from "react";
-import { Routes, Route, Link, useNavigate } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import Search from "../Components/SearchBar.jsx";
-import Plot from "../Pages/Plot.jsx";
-import About from "../Pages/About.jsx";
 import NavBar from "../Components/NavBar.jsx";
 import LoadingGif from "../assets/loadingGif.gif";
-import ErrorPage from "../Pages/ErrorPage.jsx";
 import { VscStarHalf } from "react-icons/vsc";
+import { FaArrowUp } from "react-icons/fa";
 import Footer from "../Components/Footer.jsx";
-import Login from "./Login";
-import SignUp from "../Pages/SignUp.jsx";
-import Upcoming from "../Components/Trending.jsx";
+import Trending from "../Components/Trending.jsx";
 import useFetch from "../Hooks/useFetch.jsx";
-import TrendingMovies from "../Components/Trending.jsx";
-import Welcome from "./Welcome.jsx";
+import UpComingMovies from "../Components/UpComing.jsx";
+import Trailers from "../Components/Trailers.jsx";
+import "../assets/Home.css"; // Import dark theme CSS
 
-const TMDB_BASE_URL = "https://api.themoviedb.org/3";
+const URL = "https://api.themoviedb.org/3";
 
-function Home() {
+function Home({
+  apiKey,
+  user = { username: "Guest" },
+  onLogout = () => {},
+  defaultSearchQuery = "",
+}) {
   const [url, setUrl] = useState("");
-  const [searchTerm, setSearchTerm] = useState("");
+  const [searchTerm, setSearchTerm] = useState(defaultSearchQuery);
   const { data: movies, loading, error } = useFetch(url);
-  const [userName, setUserName] = useState("");
+  const [userName, setUserName] = useState(user.username);
+  const [showScrollTop, setShowScrollTop] = useState(false);
   const navigate = useNavigate();
-  
 
+  const name = userName.toUpperCase();
+
+  function customTrimTitle(title) {
+    const parts = title.split('-');
+    const firstPart = parts[0]?.trim();
+    const secondPart = parts[1] ? "- " + parts[1].trim() : '';
+    return { firstPart, secondPart };
+  }
+
+  // Authentication check
   useEffect(() => {
-    // Check if the user is authenticated
     const isAuthenticated = localStorage.getItem("isAuthenticated") === "true";
     const currentUser = JSON.parse(localStorage.getItem("currentUser"));
 
     if (!isAuthenticated) {
-      navigate("/login"); // Redirect to login if not authenticated
+      navigate("/login");
     } else {
-      setUserName(currentUser?.username || "Guest");
+      setUserName(currentUser?.username || user.username);
     }
-  }, [navigate]);
+  }, [navigate, user.username]);
+
+  // Scroll event listener
+  useEffect(() => {
+    const handleScroll = () => {
+      setShowScrollTop(window.scrollY > 300);
+    };
+
+    window.addEventListener("scroll", handleScroll);
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, []);
+
+  // Handle window close
+  useEffect(() => {
+    const handleBeforeUnload = (event) => {
+      localStorage.removeItem("isAuthenticated");
+      localStorage.removeItem("currentUser");
+      event.preventDefault();
+      event.returnValue = "";
+    };
+
+    window.addEventListener("beforeunload", handleBeforeUnload);
+    return () => window.removeEventListener("beforeunload", handleBeforeUnload);
+  }, []);
 
   const handleSearch = (query) => {
     setSearchTerm(query);
-    setUrl(`${TMDB_BASE_URL}/search/movie?query=${query}`);
+    setUrl(`${URL}/search/movie?api_key=${apiKey}&query=${encodeURIComponent(query)}`);
   };
+
+  const scrollToTop = () => {
+    window.scrollTo({ top: 0, behavior: "smooth" });
+  };
+
+  const handleLogout = () => {
+    onLogout(navigate);
+    localStorage.removeItem("isAuthenticated");
+    localStorage.removeItem("currentUser");
+    navigate("/login");
+  };
+
   return (
-    <>
+    <div className="min-h-screen">
       <NavBar />
-      <h2>Welcome to Movie Bluff, {userName}!</h2>
-      <Routes>
-        <Route
-          path="/home"
-          element={
-            <div className="wrapper-search">
-              <Search onSearch={handleSearch} />
+      <h1
+        className="welcome-heading"
+        onClick={handleLogout}
+        title="Click to logout"
+      >
+        Welcome, {name}!
+      </h1>
+      <div className="wrapper-search">
+        <Search onSearch={handleSearch} initialQuery={defaultSearchQuery} />
+        {loading && (
+          <div className="loading-container">
+            <img src={LoadingGif} alt="Loading..." width="200" />
+          </div>
+        )}
+        {error && <p className="error-message">{error}</p>}
 
-              {loading && (
-                <div className="loading-container">
-                  <img src={LoadingGif} alt="Loading..." width="200" />
+        {searchTerm && movies?.length > 0 && (
+          <p className="search-results">
+            {`Results for "${searchTerm}": ${movies.length}`}
+          </p>
+        )}
+        {searchTerm && movies?.length === 0 && !loading && (
+          <p className="no-results">
+            No results found for "{searchTerm}"
+          </p>
+        )}
+
+        {!searchTerm && (
+          <>
+            <Trailers />
+            <Trending />
+            <UpComingMovies />
+          </>
+        )}
+
+        {movies?.length > 0 && (
+          <div className="movie-grid">
+            {movies.map((movie) => {
+              const { firstPart, secondPart } = customTrimTitle(movie.title);
+
+              return (
+                <div className="movie-card" key={movie.id}>
+                  <h2>
+                    {firstPart}
+                    <br />
+                    <span className="subtitle">{secondPart}</span>
+                  </h2>
+                  <p className="release-year">{movie.release_date?.split("-")[0]}</p>
+                  <p className="language">Language: {movie.original_language.toUpperCase()}</p>
+                  <p className="rating">
+                    <VscStarHalf className="star-icon" /> {movie.vote_average.toFixed(1)}
+                  </p>
+                  {movie.poster_path ? (
+                    <Link to={`/plot?id=${movie.id}`}>
+                      <img
+                        src={`https://image.tmdb.org/t/p/w500/${movie.poster_path}`}
+                        alt={movie.title}
+                        className="movie-poster"
+                      />
+                    </Link>
+                  ) : (
+                    <img
+                      src="https://via.placeholder.com/500x750?text=No+Image+Available"
+                      alt={movie.title}
+                      className="movie-poster"
+                    />
+                  )}
                 </div>
-              )}
+              );
+            })}
+          </div>
+        )}
 
-              {error && <p className="error-message">{error}</p>}
+        {movies?.length === 1 && searchTerm && (
+          <div className="single-movie-info">
+            <h2>{movies[0].title}</h2>
+            <p>Language: {movies[0].original_language.toUpperCase()}</p>
+            <p>Release Date: {movies[0].release_date}</p>
+            <p>Rating: {movies[0].vote_average.toFixed(1)}</p>
+            <p>Vote Count: {movies[0].vote_count}</p>
+            {movies[0].poster_path && (
+              <img
+                src={`https://image.tmdb.org/t/p/w500/${movies[0].poster_path}`}
+                alt={movies[0].title}
+                className="movie-poster"
+              />
+            )}
+          </div>
+        )}
+      </div>
 
-              {searchTerm && movies && movies.length > 0 && (
-                <p className="results-info">
-                  {`Results for "${searchTerm}": ${movies.length}`}
-                </p>
-              )}
+      {showScrollTop && (
+        <button
+          onClick={scrollToTop}
+          className="scroll-top-button"
+          aria-label="Scroll to top"
+        >
+          <FaArrowUp size={20} />
+        </button>
+      )}
 
-              {!searchTerm || (movies && movies.length === 0) ? <Upcoming /> : null} 
-
-              {movies && movies.length > 0 && (
-                <div className="movie-grid">
-                  {movies.map((movie) => (
-                    <div key={movie.id} className="movie-card">
-                      <h2>{movie.title}</h2>
-                      <p>{movie.release_date?.split("-")[0]}</p>
-                      <p>Language: {movie.original_language.toUpperCase()}</p>
-                      <p>
-                        <VscStarHalf /> {movie.vote_average.toFixed(1)}
-                      </p>
-
-                      <Link to={`/plot?id=${movie.id}`}>
-                        <img
-                          src={`https://image.tmdb.org/t/p/w500/${movie.poster_path}`}
-                          alt={movie.title}
-                          className="movie-poster"
-                        />
-                      </Link>
-                    </div>
-                  ))}
-                </div>
-              )}
-              <Footer />
-            </div>
-          }
-        />
-        <Route path="/plot" element={<Plot />} />
-        <Route path="/welcome" element={<Welcome />} />
-        <Route path="/trending" element={<TrendingMovies />} />
-        <Route path="/about" element={<About />} />
-        <Route path="/login" element={<Login />} />
-        <Route path="/signup" element={<SignUp />} />
-        <Route path="/*" element={<ErrorPage />} />
-      </Routes>
-    </>
+      <Footer />
+    </div>
   );
 }
