+++ src/assets/MovieReview.css
@@ -1,5 +1,4 @@
 /* src/assets/MovieReview.css */
-
 .movie-reviews {
   max-width: 800px;
   margin: 20px auto;
