+++ src/assets/UpComing.css
@@ -1,59 +1,52 @@
 /* Main container */
 .upcoming-container {
-  padding: 20px; /* Reduced for mobile */
+  padding: 40px;
   background: white;
   text-align: left;
   border-radius: 10px;
-  display: flex; /* Changed to flex for better alignment */
-  flex-direction: column; /* Stack children vertically */
-  width: 100%; /* Full width */
-  box-sizing: border-box; /* Include padding in width */
+  display: contents;
 }
 
 /* Title styling */
 .upcoming-title {
-  font-size: 18px; /* Slightly smaller for mobile */
-  padding-left: 15px; /* Adjusted for smaller screens */
+  font-size: 20px;
+  padding-left: 20px;
   color: #222;
-  margin-bottom: 15px; /* Reduced margin */
+  margin-bottom: 20px;
   font-weight: bold;
 }
 
 /* Scrollable movie grid (right to left) */
 .upcoming-movie-grid {
   display: flex;
-  flex-direction: row;
-  gap: 10px; /* Smaller gap for mobile */
-  padding: 10px; /* Reduced padding */
+  flex-direction: row; 
+  gap: 15px;
+  padding: 15px;
   overflow-x: auto;
   overflow-y: hidden;
   white-space: nowrap;
   scroll-behavior: smooth;
   border-radius: 10px;
-  box-sizing: border-box;
-  flex-wrap: nowrap; /* Ensure single row */
-  -webkit-overflow-scrolling: touch; /* Smooth scrolling on mobile */
+  padding-bottom: 60px;
 }
 
 /* Hide scrollbar */
 .upcoming-movie-grid::-webkit-scrollbar {
-  display: none; /* Cleaner look */
+  display:flex;
 }
 
 /* Movie card styling */
 .upcoming-movie-card {
   background: white;
   border-radius: 12px;
-  padding: 10px; /* Reduced for mobile */
+  padding: 15px;
   box-shadow: 0px 5px 10px rgba(0, 0, 0, 0.2);
   transition: transform 0.3s ease-in-out, box-shadow 0.3s ease-in-out;
-  flex: 0 0 auto; /* Prevent shrinking/stretching */
-  width: 160px; /* Smaller width for mobile */
-  display: flex; /* Flex container for card content */
-  flex-direction: column; /* Stack image, title, etc. */
+  min-width: fit-content; 
+  min-height: fit-content;
 }
 
-/* Hover effect (disabled on mobile, using :hover for desktop) */
+/* Hover effect */
 .upcoming-movie-card:hover {
   transform: scale(1.05);
   box-shadow: 0px 8px 16px rgba(0, 0, 0, 0.3);
@@ -62,28 +55,30 @@
 /* Movie poster */
 .upcoming-movie-image {
   width: 100%;
-  height: 240px; /* Smaller height for mobile */
+  height: 320px;
   object-fit: cover;
   border-radius: 10px;
 }
 
 /* Movie title */
 .upcoming-movie-title {
-  font-size: 16px; /* Smaller for mobile */
-  color: black;
-  margin-top: 8px;
-  transition: font-size 0.3s ease-in-out; /* Smooth title transition */
+font-size: 18px;
+color: black;
+margin-top: 10px;
+transition: transform 0.3s ease-in-out, font-size 0.3s ease-in-out;
 }
 
-/* Enlarge title on hover (desktop only) */
-.upcoming-movie-card:hover .upcoming-movie-title {
-  font-size: 18px;
-  font-weight: bold;
+/* Enlarge only the title on hover */
+.upcoming-movie-card:hover {
+transform: scale(1.05);
+font-size: 18px;
+font-weight: bold;
 }
+
 
 /* Movie release year */
 .upcoming-movie-year {
-  font-size: 12px; /* Smaller for mobile */
+  font-size: 14px;
   color: #777;
   margin-top: 5px;
 }
@@ -92,74 +87,27 @@
 .upcoming-rating {
   color: #2ecc71;
   font-weight: bold;
-  font-size: 12px; /* Smaller for mobile */
+  font-size: 14px;
   margin-top: 5px;
 }
 
 /* Popularity */
 .upcoming-popularity {
   color: #ff5733;
-  font-size: 12px; /* Smaller for mobile */
+  font-size: 14px;
   font-weight: bold;
 }
 
 /* Error message */
 .upcoming-error {
   color: red;
-  font-size: 14px; /* Smaller for mobile */
+  font-size: 16px;
   font-weight: bold;
-  text-align: center; /* Centered for visibility */
 }
 
 /* Loading message */
 .upcoming-loading {
-  font-size: 16px; /* Smaller for mobile */
+  font-size: 18px;
   font-weight: bold;
   color: blue;
-  text-align: center; /* Centered for visibility */
-}
-
-/* Media query for larger screens (desktop) */
-@media (min-width: 768px) {
-  .upcoming-container {
-    padding: 40px; /* Restore original padding */
-  }
-
-  .upcoming-title {
-    font-size: 20px;
-    padding-left: 20px;
-    margin-bottom: 20px;
-  }
-
-  .upcoming-movie-grid {
-    gap: 15px;
-    padding: 15px;
-  }
-
-  .upcoming-movie-card {
-    padding: 15px;
-    width: 200px; /* Larger card width */
-  }
-
-  .upcoming-movie-image {
-    height: 320px; /* Restore original height */
-  }
-
-  .upcoming-movie-title {
-    font-size: 18px;
-  }
-
-  .upcoming-movie-year,
-  .upcoming-rating,
-  .upcoming-popularity {
-    font-size: 14px;
-  }
-
-  .upcoming-error {
-    font-size: 16px;
-  }
-
-  .upcoming-loading {
-    font-size: 18px;
-  }
 }
